+++ templates/tutor/extended/escaped_customised.py
@@ -217,6 +217,11 @@ def lookup(key):
                     "unique_ender": "@",
                     "prefixes": {
                         "#": "escape"
+                        },
+                    "numbers": {
+                        "-S": "2",
+                        "-T": "3",
+                        "-TS": "4",
                         }
                     }
                 )